requirements.txt
aioredis==1.3.1
amqp==5.0.2
arabic-reshaper==2.1.3
arrow==1.2.3
asgiref==3.6.0
asn1crypto==1.5.1
astroid==2.14.1
async-timeout==4.0.2
asyncio==3.4.3
attrs==22.1.0
autobahn==23.1.2
Automat==22.10.0
beautifulsoup4==4.11.1
billiard==3.6.3.0
bson==0.5.10
celery==5.0.4
certifi==2022.9.24
cffi==1.14.4
channels==3.0.5
channels-redis==4.1.0
charset-normalizer==2.1.1
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
colorama==0.4.6
constantly==15.1.0
cryptography==40.0.1
cssselect2==0.6.0
daphne==3.0.2
Deprecated==1.2.13
dill==0.3.6
Django==4.1.3
django-ckeditor==6.4.2
django-el-pagination==4.0.0
django-environ==0.9.0
django-extensions==3.2.1
django-jazzmin==2.6.0
django-js-asset==2.0.0
django-material==1.11.3
django-rest-framework==0.1.0
django-user-agents==0.4.0
djangorestframework==3.13.1
future==0.18.2
hiredis==1.1.0
html5lib==1.1
htmldocx==0.0.6
hyperlink==21.0.0
idna==3.4
importlib-metadata==4.8.3
importlib-resources==5.4.0
incremental==22.10.0
isort==5.12.0
joblib==1.1.0
jwcrypto==1.0
jwt==1.3.1
kombu==5.0.2
lazy-object-proxy==1.9.0
lxml==4.9.2
MarkupSafe==2.1.1
marshmallow==3.19.0
mccabe==0.7.0
msgpack==1.0.3
nltk==3.6.7
numpy==1.24.2
oscrypto==1.3.0
packaging==21.3
Pillow==9.3.0
platformdirs==3.0.0
prompt-toolkit==3.0.8
psycopg2-binary==2.9.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyHamcrest==2.0.2
pyHanko==0.13.2
pyhanko-certvalidator==0.19.5
pyOpenSSL==23.1.1
pypandoc==1.8.1
pyparsing==3.0.9
PyPDF2==3.0.1
PyPDF3==1.0.6
pypng==0.20220715.0
python-bidi==0.4.2
python-dateutil==2.8.2
python-docx==0.8.11
python-jwt==3.3.2
pytz==2022.6
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
qrcode==7.3.1
redis==4.5.4
regex==2022.3.15
reportlab==3.6.12
requests==2.28.1
service-identity==18.1.0
six==1.16.0
soupsieve==2.3.2.post1
sqlparse==0.4.3
svglib==1.4.1
tinycss2==1.1.1
tomli==2.0.1
tomlkit==0.11.6
tqdm==4.63.1
Twisted==22.10.0
txaio==23.1.1
typing_extensions==4.4.0
tzdata==2022.6
tzlocal==4.2
ua-parser==0.16.1
uritools==4.0.0
urllib3==1.26.13
user-agents==2.2.0
vine==5.0.0
wcwidth==0.2.5
webencodings==0.5.1
wrapt==1.14.1
WTForms==3.0.1
xhtml2pdf==0.2.8
zipp==3.6.0
zope.interface==5.5.2
